Make vtable peek/poke 64-bit-safe (pointer hygiene, stage 3j)
pvGetVTable/PokeVTable read and write an object's vtable pointer by
aliasing the object as uint32 - they literally Assert(sizeof(void*) ==
sizeof(uint32)). On x64 the vtable pointer is 8 bytes, so pvGetVTable
returns only its low 32 bits (a truncated, wrong pointer) and PokeVTable
writes only the low half, leaving the high 32 bits stale. Partition.cpp's
structure-validity check then dereferences that truncated pointer, which
crashes on x64. Alias through uintptr instead so the full vtable pointer
is read/written; the asserts become sizeof(void*) == sizeof(uintptr)
(always true) and the "32 bit values" comments are corrected.

Found by a tree-wide sizeof(pointer) sweep. Two nearby findings were NOT
bugs and are left alone: DepthSort.cpp:3367's `* sizeof(CRenderPolygon*)
/ 4` divides by the loop stride (i += 4), not a hardcoded pointer size,
and it scales correctly; the many `new/_alloca/memset` sites that
allocate arrays of pointers with sizeof(T*) already scale with pointer
size. The GROFF/.tpa serialization that writes sizeof(pointer)-sized
slots (GroffIO.cpp, FileIO.cpp) is the separate binary-format compat
problem, not touched here.

Behaviour-identical on Win32 where uintptr == uint32: built green (exit 0,
no new warnings), smoke-launched OK. Tenth slice of the uint32->uintptr
sweep.

+++ jp2_pc/Source/Lib/Std/StdLibEx.hpp
@@ -982,13 +982,13 @@ inline void* pvGetVTable
 //**************************************
 {
 	Assert(pv_object);
-	Assert(sizeof(void*) == sizeof(uint32));
+	Assert(sizeof(void*) == sizeof(uintptr));
 
 	//
-	// Recast as an array of 32 bit values, get the top value and cast it as a pointer to
+	// Recast as an array of pointer-width values, get the top value and cast it as a pointer to
 	// a type void.
 	//
-	return (void*)(*((uint32*)pv_object));
+	return (void*)(*((uintptr*)pv_object));
 };
 
 //*********************************************************************************************
@@ -1010,10 +1010,10 @@ inline void PokeVTable
 {
 	Assert(pv_object);
 	Assert(pv_table);
-	Assert(sizeof(void*) == sizeof(uint32));
+	Assert(sizeof(void*) == sizeof(uintptr));
 
-	// Recast as an array of 32 bit values and poke the top value.
-	*((uint32*)pv_object) = uint32(pv_table);
+	// Recast as an array of pointer-width values and poke the top value.
+	*((uintptr*)pv_object) = (uintptr)pv_table;
 };
 
 //
